Feature/fe cashout integrate

+++ server/services/cashoutService.js
@@ -1,36 +1,11 @@
+const { Op } = require("sequelize");
 const db = require("../models/index.js");
-const { createStripeAccountId, updateStripeAccount, retrieveAccountStripe, createStripeLinkVerify, createPayout, retrievePayout, deleteWithdrawMethod } = require("./stripeService.js");
+const { createStripeAccountId, updateStripeAccount, retrieveAccountStripe, createStripeLinkVerify, createPayout, retrievePayout, deleteWithdrawMethod, createNewBankAccount } = require("./stripeService.js");
 const { Channel, User, WithdrawInfor, Withdraw, sequelize } = db;
 
 
-const createMethodWithdraw = async(channelId, bankData) => {
-  const bankNumberRegex = /^\d{9,16}$/;
-  const routingNumberRegex = /^\d{4}-\d{3}$/;
+const createMethodWithdraw = async(channelId, bankToken) => {
   try {
-    if(!bankData.bankHolderName || !bankData.bankNumber || !bankData.routingNumber){
-      return {
-        status: 400,
-        data: null,
-        message: "Bank data not null."
-      }
-    }
-
-    // Validate Bank Singapore
-    if (!bankNumberRegex.test(bankData.bankNumber)) {
-      return {
-        status: 400,
-        data: null,
-        message: "Bank number must be 9-16 digits."
-      }
-    }
-
-    if (!routingNumberRegex.test(bankData.routingNumber)) {
-      return {
-        status: 400,
-        data: null,
-        message: "Routing number must be in the format XXXX-XXX."
-      }
-    }
 
     const channel = await Channel.findOne({
       where: {
@@ -48,25 +23,46 @@ const createMethodWithdraw = async(channelId, bankData) => {
       const account = await createStripeAccountId(channel.User.email)
       channel.stripeAccountId = account
       await channel.save()
+
+      const res = await updateStripeAccount(channel.stripeAccountId, channel.User, bankToken)
+
+      const createWithdrawInfor = await WithdrawInfor.create({
+        channelId: channelId,
+        bankName: res.external_accounts.data[0].bank_name,
+        bankHolderName: res.external_accounts.data[0].account_holder_name,
+        bankNumber: res.external_accounts.data[0].last4,
+        routingNumber: res.external_accounts.data[0].routing_number,
+        stripeBankId: res.external_accounts.data[0].id,
+        status: res.requirements.currently_due[0] || "verified"
+      })
+
+      return {
+        status: 200,
+        data: createWithdrawInfor,
+        message: "Successful"
+      }
+    } else {
+      const res = await createNewBankAccount(channel.stripeAccountId, bankToken)
+
+      const createWithdrawInfor = await WithdrawInfor.create({
+        channelId: channelId,
+        bankName: res.bank_name,
+        bankHolderName: res.account_holder_name,
+        bankNumber: res.last4,
+        routingNumber: res.routing_number,
+        stripeBankId: res.id,
+        status: "verified"
+      })
+
+      // await detachBankAccount(channelId, channel.stripeAccountId)
+      return {
+        status: 200,
+        data: createWithdrawInfor,
+        message: "Successful"
+      }
     }
 
-    const res = await updateStripeAccount(channel.stripeAccountId, channel.User, bankData)
 
-    const createWithdrawInfor = await WithdrawInfor.create({
-      channelId: channelId,
-      bankName: res.external_accounts.data[0].bank_name,
-      bankHolderName: bankData.bankHolderName,
-      bankNumber: res.external_accounts.data[0].last4,
-      routingNumber: bankData.routingNumber,
-      stripeBankId: res.external_accounts.data[0].id,
-      status: res.requirements.currently_due[0] || "verified"
-    })
-
-    return {
-      status: 200,
-      data: createWithdrawInfor,
-      message: "Successful"
-    }
   } catch (error) {
     return {
       status: 500,
@@ -151,9 +147,20 @@ const cashout = async(channelId, repInput) => {
       },
       include: [{
         model: WithdrawInfor,
-        attributes: ['stripeBankId']
+        attributes: ['stripeBankId'],
+        where: {
+          status: 'verified'
+        }
       }]
     })
+
+    if (!channel.WithdrawInfors.stripeBankId) {
+      return {
+        status: 400,
+        data: null,
+        message: 'Insert your bank account',
+      };
+    }
 
     if (repInput < 2500) {
       return {
@@ -185,7 +192,12 @@ const cashout = async(channelId, repInput) => {
       }
     )
 
-    await channel.decrement('rep', { by: repInput });
+    await Channel.decrement('rep', { 
+      by: repInput,
+      where: {
+        id: channelId
+      }
+    });
 
     // chưa xử lý trừ rep nè check nghe
 
@@ -244,14 +256,27 @@ const checkStatusStripePayout = async (channelId) => {
 };
 
 
-const getListCashoutHistory = async(channelId) => {
+const getListCashoutHistory = async(channelId, page, pageSize, startDate, endDate, sortCondition) => {
   try {
     await checkStatusStripePayout(channelId);
 
-    const listCashout = await Withdraw.findAll({
-      where: {
-        channelId: channelId,
+    const whereCondition = { channelId };
+
+    if (startDate && endDate) {
+      if (!endDate.includes(' ') && !startDate.includes(' ')) {
+        endDate = `${endDate} 23:59:59`;
+        startDate = `${startDate} 00:00:00`;
       }
+      whereCondition.createdAt = {
+          [Op.between]: [startDate, endDate]
+      };
+    }
+
+    const listCashout = await Withdraw.findAll({
+      where: whereCondition,
+      order: [[sortCondition.sortBy, sortCondition.order]],
+      offset: (page - 1) * pageSize *1,
+      limit: pageSize*1
     })
 
     return {
@@ -270,7 +295,7 @@ const getListCashoutHistory = async(channelId) => {
 
 const getWithdrawInfor = async(channelId) => {
   try {
-    const withdrawInfor = await WithdrawInfor.findOne({where: { channelId: channelId }})
+    const withdrawInfor = await WithdrawInfor.findOne({where: { channelId: channelId, status: { [Op.ne]: 'deleted'} }})
 
     return {
       status: 200,
@@ -286,23 +311,67 @@ const getWithdrawInfor = async(channelId) => {
   }
 }
 
+const detachBankAccount = async (channelId, accountId) => {
+  try {
+    const listDelete = await WithdrawInfor.findAll({
+      where: { channelId: channelId, status: 'deleted' },
+    });
+
+    if (!listDelete.length) {
+      console.log('No records with status "deleted" found.');
+      return;
+    }
+
+    for (const withdrawInfor of listDelete) {
+      try {
+
+        await deleteWithdrawMethod(accountId, withdrawInfor.stripeBankId);
+        console.log(`Successfully deleted withdraw method: ${withdrawInfor.stripeBankId}`);
+
+        await WithdrawInfor.destroy({
+          where: {
+            channelId: channelId,
+            status: 'deleted',
+            stripeBankId: withdrawInfor.stripeBankId,
+          },
+        });
+
+      } catch (err) {
+        console.error(`Error processing stripeBankId ${withdrawInfor.stripeBankId}: ${err.message}`);
+      }
+    }
+
+    console.log('All deletions processed successfully.');
+  } catch (error) {
+    console.error('Error in detachBankAccount:', error.message);
+  }
+};
+
+
 const deleteWithdrawInfor = async(channelId, stripeBankId) => {
   try {
-    const channel = await Channel.findOne({where: { id: channelId }})
-
-    await deleteWithdrawMethod(channel.stripeAccountId, stripeBankId)
-
-    await WithdrawInfor.destroy({
-      where: {
-        channelId: channelId,
-        stripeBankId: stripeBankId
+    const [updatedRows] = await WithdrawInfor.update(
+      { status: 'deleted' },
+      {
+        where: {
+          channelId: channelId,
+          stripeBankId: stripeBankId,
+        },
       }
-    })
+    );
 
-    return {
-      status: 200,
-      message: "Bank removed successfully"
+    if (updatedRows === 0) {
+      return {
+        status: 404,
+        message: "No records were removed."
+      }
+    } else {
+      return {
+        status: 200,
+        message: "Bank removed successfully"
+      }
     }
+
 
   } catch (error) {
     return {
